+++ custom-recipes/pi-system-retrieve-event-frames/recipe.py
@@ -4,7 +4,10 @@ import copy
 from dataiku.customrecipe import get_input_names_for_role, get_recipe_config, get_output_names_for_role
 import pandas as pd
 from safe_logger import SafeLogger
-from osisoft_plugin_common import get_credentials, get_interpolated_parameters, get_advanced_parameters, check_debug_mode, get_max_count
+from osisoft_plugin_common import (
+    get_credentials, get_interpolated_parameters,
+    get_advanced_parameters, check_debug_mode, PerformanceTimer, get_max_count, reorder_dataframe
+)
 from osisoft_constants import OSIsoftConstants
 from osisoft_client import OSIsoftClient
 
@@ -41,8 +44,13 @@ start_time_column = config.get("start_time_column")
 use_end_time_column = config.get("use_end_time_column", False)
 end_time_column = config.get("end_time_column")
 server_url_column = config.get("server_url_column")
+search_full_hierarchy = config.get("search_full_hierarchy", None)
 use_batch_mode, batch_size = get_advanced_parameters(config)
 interval, sync_time, boundary_type = get_interpolated_parameters(config)
+
+network_timer = PerformanceTimer()
+processing_timer = PerformanceTimer()
+processing_timer.start()
 
 input_parameters_dataset = dataiku.Dataset(input_dataset[0])
 output_dataset = dataiku.Dataset(output_names_stats[0])
@@ -64,9 +72,15 @@ with output_dataset.get_writer() as writer:
         start_time = input_parameters_row.get(start_time_column, start_time) if use_start_time_column else start_time
         end_time = input_parameters_row.get(end_time_column, end_time) if use_end_time_column else end_time
         event_frame_webid = input_parameters_row.get("WebId")
+        event_frame_start_time = input_parameters_row.get("StartTime", "")
+        event_frame_end_time = input_parameters_row.get("EndTime", "")
 
         if client is None or previous_server_url != server_url:
-            client = OSIsoftClient(server_url, auth_type, username, password, is_ssl_check_disabled=is_ssl_check_disabled, is_debug_mode=is_debug_mode)
+            client = OSIsoftClient(
+                server_url, auth_type, username, password,
+                is_ssl_check_disabled=is_ssl_check_disabled, is_debug_mode=is_debug_mode,
+                network_timer=network_timer
+            )
             previous_server_url = server_url
         object_id = input_parameters_row.get(path_column)
         item = None
@@ -81,15 +95,17 @@ with output_dataset.get_writer() as writer:
                 interval=interval,
                 sync_time=sync_time,
                 boundary_type=boundary_type,
+                search_full_hierarchy=search_full_hierarchy,
                 can_raise=False,
                 max_count=max_count
             )
         elif use_batch_mode:
-            buffer.append(object_id)
+            buffer.append({"WebId": object_id, "StartTime": event_frame_start_time, "EndTime": event_frame_end_time})
             batch_buffer_size += 1
             if (batch_buffer_size >= batch_size) or (absolute_index == nb_rows_to_process):
                 rows = client.get_rows_from_webids(
                     buffer, data_type, max_count,
+                    search_full_hierarchy=search_full_hierarchy,
                     can_raise=False,
                     batch_size=batch_size
                 )
@@ -106,6 +122,7 @@ with output_dataset.get_writer() as writer:
                 interval=interval,
                 sync_time=sync_time,
                 boundary_type=boundary_type,
+                search_full_hierarchy=search_full_hierarchy,
                 max_count=max_count,
                 can_raise=False
             )
@@ -113,6 +130,10 @@ with output_dataset.get_writer() as writer:
         row_count = 0
         for row in rows:
             row_count += 1
+            if not use_batch_mode and event_frame_start_time:
+                row["StartTime"] = event_frame_start_time
+            if not use_batch_mode and event_frame_end_time:
+                row["EndTime"] = event_frame_end_time
             base_row = copy.deepcopy(row)
             base_row.pop("Links", None)
             items_column = base_row.pop(OSIsoftConstants.API_ITEM_KEY, [])
@@ -134,7 +155,12 @@ with output_dataset.get_writer() as writer:
                 item_row.update(base_row)
                 unnested_items_rows.append(item_row)
         unnested_items_rows = pd.DataFrame(unnested_items_rows)
+        unnested_items_rows = reorder_dataframe(unnested_items_rows, ['Path', 'Name', 'StartTime', 'EndTime', 'Timestamp', 'Value', 'UnitsAbbreviation', 'Errors'])
         if first_dataframe:
             output_dataset.write_schema_from_dataframe(unnested_items_rows)
             first_dataframe = False
         writer.write_dataframe(unnested_items_rows)
+
+processing_timer.stop()
+logger.info("Overall timer:{}".format(processing_timer.get_report()))
+logger.info("Network timer:{}".format(network_timer.get_report()))