+++ python-connectors/pi-system_event-frames-search/connector.py
@@ -7,7 +7,7 @@ from safe_logger import SafeLogger
 from osisoft_plugin_common import (
     PISystemConnectorError, RecordsLimit, get_credentials,
     build_requests_params, assert_time_format, get_advanced_parameters, check_debug_mode,
-    get_max_count
+    PerformanceTimer, get_max_count
 )
 
 
@@ -28,13 +28,20 @@ class OSIsoftConnector(Connector):
 
         auth_type, username, password, server_url, is_ssl_check_disabled = get_credentials(config)
         is_debug_mode = check_debug_mode(config)
-        self.client = OSIsoftClient(server_url, auth_type, username, password, is_ssl_check_disabled=is_ssl_check_disabled, is_debug_mode=is_debug_mode)
+        self.network_timer = PerformanceTimer()
+        self.yields_timer = PerformanceTimer()
+        self.client = OSIsoftClient(
+            server_url, auth_type, username, password,
+            is_ssl_check_disabled=is_ssl_check_disabled,
+            is_debug_mode=is_debug_mode, network_timer=self.network_timer
+        )
         self.object_id = config.get("event_frame_to_retrieve", None)
         self.data_type = config.get("data_type", "SummaryData")
         if self.object_id is None:
             self.object_id = config.get("next_element", None)
         self.start_time = config.get("start_time")
         self.end_time = config.get("end_time")
+        self.search_mode = config.get("search_mode", None)
         self.output_type = config.get("output_type")
         assert_time_format(self.start_time, error_source="start time")
         assert_time_format(self.end_time, error_source="end time")
@@ -42,6 +49,9 @@ class OSIsoftConnector(Connector):
         if not self.database_endpoint:
             raise PISystemConnectorError("No endpoint selected")
         self.must_retrieve_metrics = config.get("must_retrieve_metrics")
+        self.search_full_hierarchy = None
+        if self.must_retrieve_metrics:
+            self.search_full_hierarchy = config.get("search_full_hierarchy", None)
         self.data_type = config.get("data_type", "Recorded")
         self.max_count = get_max_count(config)
         self.config = config
@@ -63,7 +73,9 @@ class OSIsoftConnector(Connector):
             for event_frame in self.client.get_row_from_urls(
                         self.object_id, self.data_type, start_date=self.start_time,
                         end_date=self.end_time, max_count=self.max_count):
+                self.yields_timer.start()
                 yield event_frame
+                self.yields_timer.stop()
                 if limit.is_reached():
                     return
         else:
@@ -91,6 +103,7 @@ class OSIsoftConnector(Connector):
                     if use_batch_mode:
                         batch_rows = self.client.get_rows_from_webids(
                                 event_frames, self.data_type,
+                                search_full_hierarchy=self.search_full_hierarchy,
                                 can_raise=False,
                                 batch_size=self.batch_size,
                                 max_count=self.max_count
@@ -105,20 +118,25 @@ class OSIsoftConnector(Connector):
                                     if isinstance(batch_row.get("Value"), dict):
                                         value = batch_row.pop("Value", {})
                                         batch_row.update(value)
+                                    self.yields_timer.start()
                                     yield batch_row
+                                    self.yields_timer.stop()
                                     if limit.is_reached():
                                         return
                             else:
                                 batch_row.pop("Links", None)
                                 batch_row.update(value)
+                                self.yields_timer.start()
                                 yield batch_row
+                                self.yields_timer.stop()
                                 if limit.is_reached():
                                     return
                     else:
                         for event_frame in event_frames:
                             event_frame_id = event_frame.get("WebId")
                             event_frame_metrics = self.client.get_row_from_webid(
-                                event_frame_id, self.data_type, max_count=self.max_count,
+                                event_frame_id, self.data_type,
+                                search_full_hierarchy=self.search_full_hierarchy, max_count=self.max_count,
                                 can_raise=False
                             )
                             for event_frame_metric in event_frame_metrics:
@@ -136,7 +154,9 @@ class OSIsoftConnector(Connector):
                                             value = row.pop("Value", {})
                                             row.update(value)
                                         row["event_frame_webid"] = event_frame_id
+                                        self.yields_timer.start()
                                         yield row
+                                        self.yields_timer.stop()
                                         if limit.is_reached():
                                             return
                                 else:
@@ -144,7 +164,9 @@ class OSIsoftConnector(Connector):
                                     value = event_frame_copy.pop("Value", {})
                                     event_frame_copy.update(value)
                                     event_frame_copy["event_frame_webid"] = event_frame_id
+                                    self.yields_timer.start()
                                     yield event_frame_copy
+                                    self.yields_timer.stop()
                                     if limit.is_reached():
                                         return
                 else:
@@ -153,12 +175,16 @@ class OSIsoftConnector(Connector):
                         event_frame_copy.pop(OSIsoftConstants.API_ITEM_KEY, None)
                         event_frame_copy.pop("Security", None)
                         event_frame_copy.pop("Links", None)
+                        self.yields_timer.start()
                         yield event_frame_copy
+                        self.yields_timer.stop()
                         if limit.is_reached():
                             return
         end_time = datetime.datetime.now()
         duration = end_time - start_time
         logger.info("generate_rows overall duration = {}s".format(duration.microseconds/1000000 + duration.seconds))
+        logger.info("Network timer:{}".format(self.network_timer.get_report()))
+        logger.info("DSS ingress:{}".format(self.yields_timer.get_report()))
 
     def get_writer(self, dataset_schema=None, dataset_partitioning=None,
                    partition_id=None):