+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 - Fix Toolbox issue with tags and paths containing spaces
 - Fix Toolbox result consistency when using web IDs.
 - Add option to change max count value
+- Add performance timers
+- Add option to parse object columns as string
+- Transpose & Synchronization recipe output made DB friendly
+- Add Event Frames Search Full Hierarchy
+- Add time span search modes to Event Frame search connector
+- Fix write issue when Value column contains empty cells
 
 ## [Version 1.0.1](https://github.com/dataiku/dss-plugin-pi-server/releases/tag/v1.0.1) - Feature release - 2023-01-05